refactor model layers
