+++ TTS/tts/layers/delightful_tts/energy_adaptor.py
@@ -1,13 +1,12 @@
-from typing import Callable, Tuple
+from typing import Tuple, Callable
 
 import torch
-from torch import nn
+import torch.nn as nn
 
-from TTS.tts.layers.delightful_tts.variance_predictor import VariancePredictor
 from TTS.tts.utils.helpers import average_over_durations
+from TTS.tts.layers.d_tts.variance_predictor import VariancePredictor
 
-
-class EnergyAdaptor(nn.Module):  # pylint: disable=abstract-method
+class EnergyAdaptor(nn.Module): # pylint: disable=abstract-method
     """Variance Adaptor with an added 1D conv layer. Used to
     get energy embeddings.
 
@@ -30,7 +29,6 @@ class EnergyAdaptor(nn.Module):  # pylint: disable=abstract-method
         - **average energy target(train only)** (batch, 1, time1): Tensor produced after averaging over durations
 
     """
-
     def __init__(
         self,
         channels_in: int,
